refactor(ui): move directories + fix setting imports
