--- app/src/main/java/com/nktnet/webview_kiosk/ui/components/common/settings/fields/GenericSettingFieldItem.kt
@@ -1,49 +1,0 @@
-package com.nktnet.webview_kiosk.ui.components.common.settings.fields
-
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.*
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Edit
-import androidx.compose.material3.*
-import androidx.compose.material3.DividerDefaults
-import androidx.compose.material3.HorizontalDivider
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-
-@Composable
-fun <T> GenericSettingFieldItem(
-    label: String,
-    value: T,
-    onClick: () -> Unit,
-    description: @Composable (T) -> Unit
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable { onClick() }
-            .padding(top = 8.dp, start = 2.dp, end = 2.dp)
-    ) {
-        Box(modifier = Modifier.fillMaxWidth()) {
-            Column(modifier = Modifier.fillMaxWidth(0.9f)) {
-                Text(
-                    text = label,
-                    style = MaterialTheme.typography.titleMedium
-                )
-                Spacer(modifier = Modifier.height(2.dp))
-                description(value)
-            }
-            Icon(
-                imageVector = Icons.Default.Edit,
-                contentDescription = "Edit",
-                modifier = Modifier.align(Alignment.CenterEnd)
-            )
-        }
-        HorizontalDivider(
-            modifier = Modifier.padding(top = 8.dp),
-            thickness = DividerDefaults.Thickness,
-            color = DividerDefaults.color
-        )
-    }
-}
