--- app/src/main/java/com/nktnet/webview_kiosk/ui/components/setting/SettingsActionButton.kt
@@ -1,80 +1,0 @@
-package com.nktnet.webview_kiosk.ui.components.setting
-
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedButton
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import androidx.navigation.NavController
-import com.nktnet.webview_kiosk.config.Screen
-
-@Composable
-fun SettingsActionButtons(
-    navController: NavController,
-    saveEnabled: Boolean,
-    saveSettings: () -> Unit,
-) {
-    Column(modifier = Modifier.fillMaxWidth()) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(top = 20.dp),
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            OutlinedButton(
-                onClick = { navController.popBackStack() },
-                modifier = Modifier.width(150.dp),
-                colors = ButtonDefaults.outlinedButtonColors(
-                    contentColor = MaterialTheme.colorScheme.error
-                ),
-                border = BorderStroke(1.dp, MaterialTheme.colorScheme.error)
-            ) {
-                Text("Cancel")
-            }
-
-            OutlinedButton(
-                enabled = saveEnabled,
-                modifier = Modifier.width(150.dp),
-                onClick = { saveSettings() },
-                colors = ButtonDefaults.outlinedButtonColors(
-                    contentColor = MaterialTheme.colorScheme.primary
-                ),
-                border = BorderStroke(1.dp, MaterialTheme.colorScheme.primary)
-            ) {
-                Text("Save")
-            }
-        }
-
-        Button(
-            onClick = {
-                saveSettings()
-                navController.navigate(Screen.WebView.route) {
-                    popUpTo(Screen.Settings.route) { inclusive = true }
-                }
-            },
-            modifier = Modifier
-                .padding(top = 2.dp)
-                .width(150.dp)
-                .align(Alignment.End),
-            enabled = saveEnabled,
-            colors = ButtonDefaults.buttonColors(
-                containerColor = MaterialTheme.colorScheme.primary,
-                contentColor = MaterialTheme.colorScheme.onPrimary
-            )
-        ) {
-            Text("Save & Close")
-        }
-    }
-}
