+++ app/src/main/java/com/nktnet/webview_kiosk/ui/components/setting/fielditems/webengine/EnableJavaScriptSetting.kt
@@ -1,0 +1,19 @@
+package com.nktnet.webview_kiosk.ui.components.setting.fielditems.webengine
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.platform.LocalContext
+import com.nktnet.webview_kiosk.config.UserSettings
+import com.nktnet.webview_kiosk.ui.components.setting.fields.BooleanSettingFieldItem
+
+@Composable
+fun EnableJavaScriptSetting() {
+    val context = LocalContext.current
+    val userSettings = UserSettings(context)
+
+    BooleanSettingFieldItem(
+        label = "Enable JavaScript",
+        infoText = "Allow the execution of JavaScript in web pages.",
+        initialValue = userSettings.enableJavaScript,
+        onSave = { userSettings.enableJavaScript = it }
+    )
+}
