--- app/src/main/java/com/nktnet/webview_kiosk/ui/components/setting/fields/appearance/WebViewInsetSetting.kt
@@ -1,23 +1,0 @@
-package com.nktnet.webview_kiosk.ui.components.setting.fields.appearance
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import androidx.compose.ui.platform.LocalContext
-import com.nktnet.webview_kiosk.config.UserSettings
-import com.nktnet.webview_kiosk.config.WebViewInset
-import com.nktnet.webview_kiosk.ui.components.common.settings.fields.DropdownSettingFieldItem
-
-@Composable
-fun WebViewInsetSetting() {
-    val context = LocalContext.current
-    val userSettings = remember { UserSettings(context) }
-
-    DropdownSettingFieldItem(
-        label = "WebView Insets",
-        infoText = "Select which WindowInsets the WebView should respect for padding.",
-        options = WebViewInset.entries,
-        initialValue = userSettings.webViewInset,
-        onSave = { userSettings.webViewInset = it },
-        itemText = { it.label }
-    )
-}
