--- app/src/main/java/com/nktnet/webview_kiosk/ui/components/setting/fields/webbrowsing/ClearHistoryOnHomeSetting.kt
@@ -1,20 +1,0 @@
-package com.nktnet.webview_kiosk.ui.components.setting.fields.webbrowsing
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import androidx.compose.ui.platform.LocalContext
-import com.nktnet.webview_kiosk.config.UserSettings
-import com.nktnet.webview_kiosk.ui.components.common.settings.fields.BooleanSettingFieldItem
-
-@Composable
-fun ClearHistoryOnHomeSetting() {
-    val context = LocalContext.current
-    val userSettings = remember { UserSettings(context) }
-
-    BooleanSettingFieldItem(
-        label = "Clear History on Home",
-        infoText = "Clear the browser history whenever the user triggers an action to return home.",
-        initialValue = userSettings.clearHistoryOnHome,
-        onSave = { userSettings.clearHistoryOnHome = it }
-    )
-}
