+++ app/src/main/java/com/nktnet/webview_kiosk/ui/view/SettingsAppearanceScreen.kt
@@ -8,10 +8,10 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import com.nktnet.webview_kiosk.ui.components.setting.SettingLabel
-import com.nktnet.webview_kiosk.ui.components.setting.fields.appearance.AddressBarModeSetting
-import com.nktnet.webview_kiosk.ui.components.setting.fields.appearance.BlockedMessageSetting
-import com.nktnet.webview_kiosk.ui.components.setting.fields.appearance.ThemeSetting
-import com.nktnet.webview_kiosk.ui.components.setting.fields.appearance.WebViewInsetSetting
+import com.nktnet.webview_kiosk.ui.components.setting.fielditems.appearance.AddressBarModeSetting
+import com.nktnet.webview_kiosk.ui.components.setting.fielditems.appearance.BlockedMessageSetting
+import com.nktnet.webview_kiosk.ui.components.setting.fielditems.appearance.ThemeSetting
+import com.nktnet.webview_kiosk.ui.components.setting.fielditems.appearance.WebViewInsetSetting
 import androidx.compose.runtime.MutableState
 import com.nktnet.webview_kiosk.config.option.ThemeOption
 import com.nktnet.webview_kiosk.ui.components.setting.SettingDivider
