+++ app/src/main/java/com/nktnet/webview_kiosk/ui/view/SettingsWebBrowsingScreen.kt
@@ -9,7 +9,7 @@ import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import com.nktnet.webview_kiosk.ui.components.setting.SettingDivider
 import com.nktnet.webview_kiosk.ui.components.setting.SettingLabel
-import com.nktnet.webview_kiosk.ui.components.setting.fields.webbrowsing.*
+import com.nktnet.webview_kiosk.ui.components.setting.fielditems.webbrowsing.*
 
 @Composable
 fun SettingsWebBrowsingScreen(
