+++ app/src/main/java/com/nktnet/webview_kiosk/ui/view/SettingsWebEngineScreen.kt
@@ -9,7 +9,7 @@ import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import com.nktnet.webview_kiosk.ui.components.setting.SettingDivider
 import com.nktnet.webview_kiosk.ui.components.setting.SettingLabel
-import com.nktnet.webview_kiosk.ui.components.setting.fields.webengine.*
+import com.nktnet.webview_kiosk.ui.components.setting.fielditems.webengine.*
 
 @Composable
 fun SettingsWebEngineScreen(navController: NavController) {
